multitool-template3.sh: added cmdcols_sample -- column widths adjusted
The new cmdcols_sample () implements command names in columns that works
like the perl implementation below.
- perl implementation adjusted to have 2 leading spaces like the other
  versions
- either perl or this version probably bugs w/ column widths, but that
  is not critical so it will be adjusted later.
- in addition to that, the echo ...| sed ...| columns version is demoed
- and finally the 'Usage' version -- to compare all (visually)

The hint to sed s/$/$/ is shown when output is to the tty -- and it
it shows the perl version prints trailing spaces. So did the initial
versions of the shell versions, which were pretty cleverly fixed. As
that is also something not too critical, those are to be updated later.